+++ resources/views/layouts/app.blade.php
@@ -10,8 +10,8 @@
             'metaDescription' => $metaDescription ?? config('app.slogan'),
         ])
 
-        <link rel="stylesheet" href="https://webblocksui.com/packages/webblocks/dist/webblocks-ui.css">
-        <link rel="stylesheet" href="https://webblocksui.com/packages/webblocks/dist/webblocks-icons.css">
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-ui.css">
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-icons.css">
     </head>
     <body>
         <div class="wb-section">
@@ -20,6 +20,6 @@
             </div>
         </div>
 
-        <script src="https://webblocksui.com/packages/webblocks/dist/webblocks-ui.js"></script>
+        <script src="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-ui.js"></script>
     </body>
 </html>