+++ resources/views/layouts/guest.blade.php
@@ -14,8 +14,8 @@
             'metaDescription' => $metaDescription ?? config('app.slogan'),
         ])
 
-        <link rel="stylesheet" href="https://webblocksui.com/packages/webblocks/dist/webblocks-ui.css">
-        <link rel="stylesheet" href="https://webblocksui.com/packages/webblocks/dist/webblocks-icons.css">
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-ui.css">
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-icons.css">
         @if (is_file($guestCssPath))
             <link rel="stylesheet" href="{{ asset('site/css/guest.css') }}?v={{ filemtime($guestCssPath) }}">
         @endif
@@ -25,7 +25,7 @@
             {{ $slot }}
         </main>
 
-        <script src="https://webblocksui.com/packages/webblocks/dist/webblocks-ui.js"></script>
+        <script src="https://cdn.jsdelivr.net/gh/fklavyenet/webblocks-ui@master/packages/webblocks/dist/webblocks-ui.js"></script>
         <script>
             document.addEventListener('click', function (event) {
                 var button = event.target.closest('[data-password-toggle]');